data-layer/requirements.txt
alembic==1.13.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==8.11.0
minio==7.2.3
requests==2.31.0
